// src/enums/hex.ts
export enum hex {
  H100 = 'FF',
  H99 = 'FC',
  H98 = 'FA',
  H97 = 'F7',
  H96 = 'F5',
  H95 = 'F2',
  H94 = 'F0',
  H93 = 'ED',
  H92 = 'EB',
  H91 = 'E8',
  H90 = 'E6',
  H89 = 'E3',
  H88 = 'E0',
  H87 = 'DE',
  H86 = 'DB',
  H85 = 'D9',
  H84 = 'D6',
  H83 = 'D4',
  H82 = 'D1',
  H81 = 'CF',
  H80 = 'CC',
  H79 = 'C9',
  H78 = 'C7',
  H77 = 'C4',
  H76 = 'C2',
  H75 = 'BF',
  H74 = 'BD',
  H73 = 'BA',
  H72 = 'B8',
  H71 = 'B5',
  H70 = 'B3',
  H69 = 'B0',
  H68 = 'AD',
  H67 = 'AB',
  H66 = 'A8',
  H65 = 'A6',
  H64 = 'A3',
  H63 = 'A1',
  H62 = '9E',
  H61 = '9C',
  H60 = '99',
  H59 = '96',
  H58 = '94',
  H57 = '91',
  H56 = '8F',
  H55 = '8C',
  H54 = '8A',
  H53 = '87',
  H52 = '85',
  H51 = '82',
  H50 = '80',
  H49 = '7D',
  H48 = '7A',
  H47 = '78',
  H46 = '75',
  H45 = '73',
  H44 = '70',
  H43 = '6E',
  H42 = '6B',
  H41 = '69',
  H40 = '66',
  H39 = '63',
  H38 = '61',
  H37 = '5E',
  H36 = '5C',
  H35 = '59',
  H34 = '57',
  H33 = '54',
  H32 = '52',
  H31 = '4F',
  H30 = '4D',
  H29 = '4A',
  H28 = '47',
  H27 = '45',
  H26 = '42',
  H25 = '40',
  H24 = '3D',
  H23 = '3B',
  H22 = '38',
  H21 = '36',
  H20 = '33',
  H19 = '30',
  H18 = '2E',
  H17 = '2B',
  H16 = '29',
  H15 = '26',
  H14 = '24',
  H13 = '21',
  H12 = '1F',
  H11 = '1C',
  H10 = '1A',
  H9 = '17',
  H8 = '14',
  H7 = '12',
  H6 = '0F',
  H5 = '0D',
  H4 = '0A',
  H3 = '08',
  H2 = '05',
  H1 = '03',
  H0 = '00'
}